+++ database/migrations/2027_05_21_000001_create_penyediaan_iklans_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('penyediaan_iklans', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('tender_id')->unique();
+            $table->unsignedInteger('tender_id')->unique();
             $table->json('meta')->nullable();
             $table->timestamp('submitted_at')->nullable();
             $table->timestamps();
